requirements.txt
tensorflow>=2.12
torch>=2.1
transformers>=4.35
datasets
faiss-cpu
scikit-learn
xgboost
shap
opencv-python
pydicom
numpy
pandas
matplotlib
albumentations
pytest
uvicorn
fastapi
peft  # pour LoRA (si tu utilises HF PEFT)
accelerate
